Load icons from frontend
This will make the document size smaller, as we don’t pass all the icons to the field to start with. But we should still use a Vuex store so that we only load all the icons once. Right now, we are still loading all the icons for each field.